+++ src/ec/entity.cpp
@@ -3,7 +3,7 @@
 #include "component.hpp"
 
 namespace  nova::ec {
-	void Entity::tick(const float delta_time)
+    void Entity::tick(const double delta_time)
 	{
 		for(auto* component : components)
 		{
